+++ src/magic_tiler/magic_tiler.py
@@ -5,8 +5,8 @@ import click
 
 import magic_tiler
 from magic_tiler import dtos
-from magic_tiler import subprocess_runner
 from magic_tiler import sway
+from magic_tiler.utils import subprocess_runner
 
 # maps from verbosity level to log levels
 VERBOSITY_LOG_LEVELS = {
